Address review feedback: active voice and move variable to reform file
- Update amount.yaml description to use active voice
- Move ut_hb210_s2_marriage_credit variable into reform file
- Delete separate variable file (following ut_hb210.py pattern)
