+++ policyengine_us/parameters/gov/contrib/states/ut/hb210/s2_marriage_credit/amount.yaml
@@ -1,4 +1,4 @@
-description: Utah HB 210 Substitute 2 marriage tax credit amount by filing status.
+description: Utah provides this marriage tax credit amount by filing status under HB 210 Substitute 2.
 
 SINGLE:
   2026-01-01: 0
